requirements.txt
eel==0.18.2
tk==0.1.0
pillow==12.2.0
pyinstaller==6.19.0